+++ funlib/learn/tensorflow/models/__init__.py
@@ -1,2 +1,3 @@
 from .conv4d import conv4d # noqa
 from .unet import unet, conv_pass, crop, downsample, upsample, repeat # noqa
+from .vgg import vgg
