+++ funlib/learn/tensorflow/models/utils.py
@@ -1,0 +1,14 @@
+import tensorflow as tf
+
+
+def get_number_of_tf_variables():
+    '''Returns number of trainable variables in tensorflow graph collection'''
+    total_parameters = 0
+    for variable in tf.trainable_variables():
+        # shape is an array of tf.Dimension
+        shape = variable.get_shape()
+        variable_parameters = 1
+        for dim in shape:
+            variable_parameters *= dim.value
+        total_parameters += variable_parameters
+    return total_parameters
